Image post without featured image now works
It's still using the _pa_post_thumbnail() function, which has also changed a lot and is now accepting a second parameter: image id.
The image to show is gotten from the first <img> tag within the content. that is then turned into an image id, which is fed to _pa_post_thumbnail().